% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_nodelist}
\alias{example_nodelist}
\title{Example nodelist file of a bipartite network.}
\format{
A data frame with 806 rows and 5 columns:
\describe{
  \item{Label}{Node ID of patients and symptoms.}
  \item{Cluster}{Cluster membership found by BipartiteModularityMaximization. Can be changed to any customized cluster membership.}
  \item{X}{X coordinates found by Fruchterman & Reingold layout. Can be changed to any customized coordinates. }
  \item{Y}{Y coordinates found by Fruchterman & Reingold layout. Can be changed to any customized coordinates. }
  \item{Entity}{Indicating whether a node is a patient (assigned 1) or a symptom (assigned 2).}
}
}
\usage{
example_nodelist
}
\description{
A dataset consisting of anonymized patients (n=798) and symptoms (d=8), where each patient has one or more symptoms.
}
\keyword{datasets}
