% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExplodeLayoutBasic.R
\name{get_edgelist_from_incidmat}
\alias{get_edgelist_from_incidmat}
\title{Generate edgelist from incidence matrix of a bipartite network}
\usage{
get_edgelist_from_incidmat(incidence_matrix)
}
\arguments{
\item{incidence_matrix}{A matrix where row names and column names are the node ID of a bipartite network. An element of the i-th row and j-th column of the matrix is 0 if node on row i is not connected to node on column j, and edge weight if they are connected.}
}
\value{
A dataframe with 3 columns: nodesR, nodesC, values.
}
\description{
Generate edgelist from incidence matrix of a bipartite network
}
\examples{
example_edgelist=get_edgelist_from_incidmat(example_incidmat)
}
