% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExplodeLayoutBasic.R
\name{get_nodelist_for_plotting}
\alias{get_nodelist_for_plotting}
\title{Format a nodelist for plotting using the new coordinates found by ExplodeLayout}
\usage{
get_nodelist_for_plotting(nodelist, new_coordinates)
}
\arguments{
\item{nodelist}{A dataframe including at least 3 columns: Label, Cluster, Entity. Entity is either 1 or 2, indicating which part of the bipartite network a node is in.}

\item{new_coordinates}{A dataframe including at least 3 columns: Label, newX, newY.}
}
\value{
A dataframe including 5 columns: Label, X, Y, Color, baseShape.
}
\description{
Format a nodelist for plotting using the new coordinates found by ExplodeLayout
}
\examples{
exploded_coords=explode_coordinates(example_nodelist,radius=1.2)
plotting_nodelist=get_nodelist_for_plotting(example_nodelist,exploded_coords)
}
